requirements.txt
mlx-whisper==0.4.2
mlx-embeddings==0.0.3